@@ -1,37 +1,37 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using SimpleRPGServer.Models;
-using SimpleRPGServer.Models.Ingame;
+using SimpleRPGServer.Persistence.Models;
+using SimpleRPGServer.Persistence.Models.Ingame;
 using SimpleRPGServer.Util;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace SimpleRPGServer.Controllers
+namespace SimpleRPGServer.Controllers;
+
+[ApiController]
+[Route("api/gamedata/ability_queue")]
+public class PlayerAbilityQueueController : ControllerBase
 {
-    [ApiController]
-    [Route("api/gamedata/ability_queue")]
-    public class PlayerAbilityQueueController : ControllerBase
+    private readonly GameDbContext _context;
+
+    public PlayerAbilityQueueController(GameDbContext context)
     {
-        private readonly GameDbContext _context;
+        this._context = context;
+    }
 
-        public PlayerAbilityQueueController(GameDbContext context)
-        {
-            this._context = context;
-        }
+    [HttpGet]
+    [Route("get")]
+    public async Task<ActionResult<List<PlayerAbility>>> GetAbilityQueue()
+    {
+        var login = HttpUtil.GetLoginFromHeader(this.Request, this._context);
+        if (login == null || login.PlayerId == 0)
+            return BadRequest();
 
-        [HttpGet]
-        [Route("get")]
-        public async Task<ActionResult<PlayerAbilityQueue>> GetAbilityQueue()
-        {
-            var login = HttpUtil.GetLoginFromHeader(this.Request, this._context);
-            if (login == null || login.PlayerId == 0)
-                return BadRequest();
+        Player player = await this._context.Players.SingleOrDefaultAsync(p => p.Id == login.PlayerId);
+        if (player == null)
+            return BadRequest();
 
-            Player player = this._context.Players.SingleOrDefault(p => p.Id == login.PlayerId);
-            if (player == null)
-                return BadRequest();
-
-            return await this._context.PlayerAbilityQueues.SingleOrDefaultAsync(paq => paq.Player.Id == player.Id);
-        }
+        return player.Abilities.Where(e => e.Queued == true).ToList();
     }
 }
